+++ NEWS.md
@@ -26,6 +26,7 @@ These changes invalidate some targets in some workflows, but they are necessary 
 
 ## Enhancements
 
+* Check for illegal formats early on at the `drake_config()` level (#1156, @MilesMcBain).
 * Smoothly deprecate the `config` argument in all user-side functions (#1118, @vkehayas). Users can now supply the plan and other `make()` arguments directly, without bothering with `drake_config()`. Now, you only need to call `drake_config()` in the `_drake.R` file for `r_make()` and friends. Old code with `config` objects should still work. Affected functions:
     * `make()`
     * `outdated()`